+++ test/src/board/coordinates.cpp
@@ -1,8 +1,8 @@
-#include <group.hpp>                            // action::is_realized, make
+#include <group.hpp>                            // action::is_realized, make_group
 #include <dctl/angle.hpp>                       // _deg, inverse, rotate
 #include <board/sequence.hpp>                   // Micro, Mini, Checkers, Roman, Spantsireti, International, Frisian, Ktar<10, 11>,
                                                 // Ktar<10, 12>, Compact1012, Compact120, Board<12, 10>, Canadian, SriLankan, Dumm
-#include <dctl/board/coordinates.hpp>            // Square, ulo_from_sq, sco_from_ulo
+#include <dctl/board/coordinates.hpp>           // operator==, rotate, to_sco
 #include <boost/iterator/counting_iterator.hpp> // counting_iterator
 #include <boost/test/test_case_template.hpp>    // BOOST_AUTO_TEST_CASE_TEMPLATE
 #include <boost/test/unit_test.hpp>             // BOOST_AUTO_TEST_SUITE, BOOST_AUTO_TEST_SUITE_END, BOOST_CHECK
@@ -45,7 +45,7 @@ BOOST_AUTO_TEST_CASE_TEMPLATE(GroupActionIsRealizedForAllCyclicGroupsOnAllSquare
         BOOST_CHECK(
                 std::all_of(begin(C_N), end(C_N), [=](auto const& g) {
                         return std::all_of(T::begin(), T::end(), [&](auto i) {
-                                auto const coord = to_sco(i, T::outer_grid);
+                                auto const coord = to_sco(static_cast<int>(i), T::outer_grid);
                                 return group::action::is_realized(coord, g);
                         });
                 })
